--- pixel-arty/INFO O BRANCHU!!!.md
@@ -1,2 +1,0 @@
-# obecnie interesujący nas model to ten z pliku pix2pix, trzeba go ztunować bo obecnie daje ubogie wyniki
-## 1. najpierw trenujemy odpalając pix2pix.py, potem możemy zrobić test z zapisanego modelu poprzez p2p_generator.py, ścieżki do plików są narazie surowe więc zmieńce na swoje
