--- pixel-arty/dataset.md
@@ -1,1 +1,0 @@
-link do datasetu: <https://drive.google.com/drive/folders/1_oQclx4JiZ5OMj4VHYmOTjGNNCAnY4uF>
