--- pixel-arty/pyproject.toml
@@ -1,14 +1,0 @@
-[project]
-name = "pixel-arty"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.12"
-dependencies = [
-	"torch",
-	"torchvision",
-	"matplotlib",
-	"numpy",
-	"pillow",
-	"dotenv"
-]
